pyproject.toml
[tool.poetry]
name = "employee-tree"
version = "0.1.0"
description = "test task"
authors = ["Asgef <asgefes1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.2"
django-bootstrap5 = "^24.3"
python-dotenv = "^1.0.1"
dj-database-url = "^2.2.0"
gunicorn = "^23.0.0"
psycopg2-binary = "^2.9.9"
django-seed = "^0.3.1"
django-filter = "^24.3"


[tool.poetry.group.dev.dependencies]
bpython = "^0.24"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
